requirements.txt
Django==2.2.8
djangorestframework==3.9.2
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
